Simplified the followers. * Removed get endpoints: followers and follows are now returned with the profile. * Moved the follow and unfollow endpoints into the students rest controller. * Updated the tests.
