+++ src/main/java/studentscroll/api/profiles/web/StudentsRestController.java
@@ -4,13 +4,17 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -19,10 +23,12 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.val;
 import studentscroll.api.account.data.Account;
 import studentscroll.api.profiles.data.Profile;
+import studentscroll.api.profiles.services.FollowersService;
 import studentscroll.api.profiles.services.ProfileService;
 import studentscroll.api.profiles.web.dto.ProfileResponse;
 import studentscroll.api.profiles.web.dto.UpdateProfileRequest;
@@ -36,6 +42,9 @@ public class StudentsRestController {
 
   @Autowired
   private ProfileService service;
+
+  @Autowired
+  private FollowersService followersService;
 
   @Operation(summary = "Find student.")
   @ApiResponses(value = {
@@ -83,6 +92,32 @@ public class StudentsRestController {
         Optional.ofNullable(request.getNewIcon()),
         Optional.ofNullable(request.getNewInterests()),
         Optional.ofNullable(request.getNewLocation())));
+  }
+
+  @Operation(summary = "Follow the student.")
+  @ApiResponses(value = {
+      @ApiResponse(responseCode = "201", description = "Followed the student."),
+      @ApiResponse(responseCode = "400", description = "Student cannot follow themselves.", content = @Content),
+      @ApiResponse(responseCode = "404", description = "Student does not exist.", content = @Content),
+      @ApiResponse(responseCode = "409", description = "Student is already followed.", content = @Content)
+  })
+  @SecurityRequirement(name = "token")
+  @PostMapping("/{studentId}/followers")
+  @ResponseStatus(HttpStatus.CREATED)
+  public ProfileResponse follow(@PathVariable Long studentId)
+      throws EntityNotFoundException, EntityExistsException, IllegalArgumentException, NotAuthenticatedException {
+    return new ProfileResponse(followersService.follow(getCurrentStudent().getProfile(), studentId));
+  }
+
+  @Operation(summary = "Unfollow the student.")
+  @ApiResponses(value = {
+      @ApiResponse(responseCode = "204", description = "Unfollowed the student."),
+      @ApiResponse(responseCode = "404", description = "Unfollower is not following or student does not exist.", content = @Content) })
+  @SecurityRequirement(name = "token")
+  @DeleteMapping("/{studentId}/followers")
+  @ResponseStatus(HttpStatus.NO_CONTENT)
+  public void unfollow(@PathVariable Long studentId) throws EntityNotFoundException, NotAuthenticatedException {
+    followersService.unfollow(getCurrentStudent().getProfile(), studentId);
   }
 
   private Account getCurrentStudent() throws NotAuthenticatedException {
